rft: 从 TaskData 中拆出 JsonMisc
